Finalization, 3 bugs found--

+++ ois/src/main/java/com/tim3/ois/service/ItemService.java
@@ -88,48 +88,58 @@ public class ItemService {
 
     public ResponseEntity<?> updateItem(int id,CreateNewItem newItem) throws ResourceNotFoundException {
         Item item = itemRepository.findById(id);
-        if(item==null)  return new ResponseEntity("Item not found with id : "+id,HttpStatus.NOT_FOUND);
+        if (item == null) return new ResponseEntity("Item not found with id : " + id, HttpStatus.NOT_FOUND);
         String fileName = newItem.getFile().getOriginalFilename();
+        MultipartFile file = newItem.getFile();
         if (StringUtils.isEmpty(fileName)) {
-            return new ResponseEntity("Please select a file!", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity("Something error occured!", HttpStatus.BAD_REQUEST);
         }
-        try {
-            Path deletePath= Paths.get(UPLOADED_FOLDER+item.getImagePath());
-            Files.delete(deletePath);
-            item.setDetail(newItem.getDetail());
-            item.setQuantity(newItem.getQuantity());
-            item.setName(newItem.getName());
-            item.setPrice(newItem.getPrice());
-            item.setImagePath(fileName);
-            saveUploadedFile(newItem.getFile());
-            saveItem(item);
-        } catch (IOException e) {
-            return new ResponseEntity<>("Some error occured. Failed to add item", HttpStatus.BAD_REQUEST);
+        if (file.getContentType().equals("image/png") || file.getContentType().equals("image/jpg") || file.getContentType().equals("image/jpeg") || file.getContentType().equals("image/bmp")) {
+            try {
+                Path deletePath = Paths.get(UPLOADED_FOLDER + item.getImagePath());
+                Files.delete(deletePath);
+                item.setDetail(newItem.getDetail());
+                item.setQuantity(newItem.getQuantity());
+                item.setName(newItem.getName());
+                item.setPrice(newItem.getPrice());
+                item.setImagePath(fileName);
+                saveUploadedFile(newItem.getFile());
+                saveItem(item);
+            } catch (IOException e) {
+                return new ResponseEntity<>("Some error occured. Failed to add item", HttpStatus.BAD_REQUEST);
+            }
+            return new ResponseEntity("Successed to update Item", HttpStatus.OK);
         }
-        return new ResponseEntity("Successed to update Item", HttpStatus.OK);
+        return new ResponseEntity("Error occured (file is not an image)", HttpStatus.BAD_REQUEST);
+
     }
 
     public ResponseEntity<?> storeItem(CreateNewItem newItem) {
         Item itemExist = itemRepository.findByName(newItem.getName());
         if(itemExist!=null)  return new ResponseEntity("Item already Exists with the name provided",HttpStatus.BAD_REQUEST);
         String fileName = newItem.getFile().getOriginalFilename();
+        MultipartFile file = newItem.getFile();
         if (StringUtils.isEmpty(fileName)) {
-            return new ResponseEntity("Please select a file!", HttpStatus.OK);
+            return new ResponseEntity("Something error occured!", HttpStatus.BAD_REQUEST);
         }
-        try {
-    //            item.setImage(bytes); // untuk save file / gambar ke field picture di database
-            Item item = new Item();
-            item.setDetail(newItem.getDetail());
-            item.setQuantity(newItem.getQuantity());
-            item.setName(newItem.getName());
-            item.setPrice(newItem.getPrice());
-            item.setImagePath(fileName);
-            saveUploadedFile(newItem.getFile());
-            saveItem(item);
-        } catch (IOException e) {
-            return new ResponseEntity<>("Some error occured. Failed to add item", HttpStatus.BAD_REQUEST);
+        if (file.getContentType().equals("image/png")||file.getContentType().equals("image/jpg")||file.getContentType().equals("image/jpeg")||file.getContentType().equals("image/bmp")) {
+            try {
+                //            item.setImage(bytes); // untuk save file / gambar ke field picture di database
+                Item item = new Item();
+                item.setDetail(newItem.getDetail());
+                item.setQuantity(newItem.getQuantity());
+                item.setName(newItem.getName());
+                item.setPrice(newItem.getPrice());
+                item.setImagePath(fileName);
+                saveUploadedFile(newItem.getFile());
+                saveItem(item);
+            } catch (IOException e) {
+                return new ResponseEntity<>("Some error occured. Failed to add item", HttpStatus.BAD_REQUEST);
+            }
+            return new ResponseEntity("Successed to add Item", HttpStatus.OK);
         }
-        return new ResponseEntity("Successed to add Item", HttpStatus.OK);
+        return new ResponseEntity("Error occured (file is not an image)", HttpStatus.BAD_REQUEST);
+
     }
         private void saveUploadedFile(MultipartFile file) throws IOException {
             if (!file.isEmpty()) {
